refactor(shared): extract generic rule-criteria types out of library-cleanup
Splits the rule-criteria types out of `library-cleanup.ts` into a new
neutral `rule-criteria.ts` so the upcoming Auto-Tagger feature can
consume the same DSL without depending on a cleanup-named module.

What moved
----------
- `cleanupRuleTypeSchema` → `ruleTypeSchema` (kept legacy alias `cleanupRuleTypeSchema`)
- `CleanupRuleType` → `RuleType` (kept legacy alias `CleanupRuleType`)
- `Condition`, `CompositeOperator`, `compositeOperatorSchema`, `conditionSchema`
- All `*RuleParams` schemas + types (~50 of them — age, size, rating,
  status, genre, year_range, video/audio codec, resolution, hdr_type,
  custom_format_score, runtime, release_group, seerr_*, tautulli_*,
  plex_*, jellyfin_*, file_path, tag_match, plex_collection,
  plex_label, plex_added_at, behavior-aware staleness/retention rules)
- `ruleParamSchemaMap` (write-time validation dispatch)
- `DataSourceDependency`, `ruleDataSourceMap`

What stayed in library-cleanup.ts
---------------------------------
- `cleanupActionSchema` / `CleanupAction` — cleanup-specific execution
- `baseCleanupRuleSchema` and the create/update/reorder/config schemas
- `CleanupRuleResponse`, `CleanupConfigResponse`, `CleanupApprovalResponse`,
  `CleanupLogResponse`, `CleanupFieldOptionsResponse`, `CleanupPreviewItem/Response`,
  `CleanupStatusResponse`, `CleanupExplainResult/Response`, `CleanupStatisticsResponse`
- `PrefetchHealthStatus` / `PrefetchSourceStatus`
- Approval queue types

Backward compatibility
----------------------
Legacy export names (`cleanupRuleTypeSchema`, `CleanupRuleType`) kept as
aliases pointing to the new generic names. All existing imports across
the codebase continue to resolve unchanged.

What did NOT move
-----------------
- The evaluator (`apps/api/src/lib/library-cleanup/rule-evaluators.ts`)
  stays where it is. The Auto-Tagger PR can import from there directly;
  the file move is pure code-organization churn that doesn't unblock
  anything new and would inflate this PR's surface unnecessarily.
- The criteria UI components (`condition-params-fields.tsx`, `multi-select-field.tsx`)
  same reasoning. Auto-Tagger can import from `features/library-cleanup/components/`.

Both can move in a follow-up cleanup PR if desired.

Validation
----------
- `pnpm --filter @arr/shared build` — clean, dist regenerated (352 → 355 KB
  for the additional re-exports through the rule-criteria barrel)
- `pnpm --filter @arr/api exec tsc --noEmit` — clean
- `pnpm --filter @arr/web exec tsc --noEmit` — clean
- `pnpm --filter @arr/api run test` — 1383 passed / 36 skipped (no regressions)
- `pnpm run lint` — no new findings

Refs: prep for Auto-Tagger arc (`memory/auto-tagger-arc.md`)
