+++ packages/shared/src/types/index.ts
@@ -14,6 +14,7 @@ export * from "./plex";
 export * from "./pulse";
 export * from "./queue-cleaner";
 export * from "./regex-safety";
+export * from "./rule-criteria";
 export * from "./search";
 export * from "./seerr";
 export * from "./tautulli";
